Moved recording_completed test to its own file and cleaned up unused imports
